modularize functions
