+++ src/constants.js
@@ -1,0 +1,5 @@
+'use strict';
+
+export default {
+  defaultFramerate: 25
+};
