+++ src/toSeconds.js
@@ -1,0 +1,18 @@
+'use strict';
+
+import { defaultFramerate } from './constants';
+
+const TIMECODE_REGEXP = /\d+\:\d+\:\d+\:\d+/;
+
+/* converts timecode to time */
+export default (timecode, frameRate = defaultFramerate) => {
+  if (!TIMECODE_REGEXP.test(timecode)) {
+    throw new Error('invalid timecode : ' + timecode);
+  }
+  if (typeof timecode === 'number') {
+    return timecode;
+  }
+  const tc = timecode.split(':');
+
+  return parseFloat(tc[0] * 60 * 60 + parseInt(tc[1], 10) * 60 + parseInt(tc[2], 10) + parseInt(tc[3], 10) / frameRate);
+};
